Add async games to recent games section (#105)
* add async matches to recent games

* try to fix deployment error

* remove NEXT_PUBLIC_BASE_URL usage
