+++ .devcontainer/devcontainer.json
@@ -7,7 +7,8 @@
             "extensions": [
                 "dsznajder.es7-react-js-snippets",
                 "dbaeumer.vscode-eslint",
-                "GitHub.copilot-chat"
+                "GitHub.copilot-chat",
+                "anthropic.claude-code"
             ]
         }
     }
