+++ src/app/page.module.css
@@ -444,6 +444,17 @@
   letter-spacing: 0.04em;
 }
 
+.modeBadge {
+  font-size: 0.65rem;
+  font-weight: 700;
+  letter-spacing: 0.06em;
+  text-transform: uppercase;
+  color: var(--color-text-muted);
+  border: 1px solid var(--color-border);
+  padding: 0.15rem 0.5rem;
+  border-radius: 4px;
+}
+
 .matchContent {
   display: flex;
   align-items: center;
