+++ src/components/MatchCard.tsx
@@ -2,14 +2,28 @@
 import Link from "next/link";
 import styles from "@/app/page.module.css";
 import { useRouter } from "next/navigation";
+import type { PlayMode } from "@/types";
 
-export default function MatchCard({ match, index }: { match: any; index: number }) {
+type RecentMatch = {
+  id: string;
+  mode: PlayMode;
+  timestamp: string;
+  player1: string;
+  player2: string;
+  playerOneId: string;
+  playerTwoId: string;
+  score: string;
+  winner: string;
+};
+
+export default function MatchCard({ match, index }: { match: RecentMatch; index: number }) {
   const router = useRouter();
-  
+
   return (
     <div key={index} className={styles.matchCard} onClick={() => router.push(`/match/${match.id}`)} style={{ cursor: "pointer" }}>
       <div className={styles.matchHeader}>
         <span className={styles.matchTime}>{match.timestamp}</span>
+        <span className={styles.modeBadge}>{match.mode === "async" ? "Async" : "Blitz"}</span>
       </div>
       <div className={styles.matchContent}>
         <Link href={`/profile/${match.playerOneId}`} onClick={(e) => e.stopPropagation()} className={styles.player}>
