+++ src/types/common.ts
@@ -1,12 +1,14 @@
 import { DateTime } from "neo4j-driver";
+import type { PlayMode } from "@/types";
 
 
-export type PlayerMatch = { 
+export type PlayerMatch = {
     id: string;
-    opponentId: string; 
-    opponentUsername: string; 
-    result: string; 
-    playerScore: number; 
-    opponentScore: number; 
+    mode: PlayMode;
+    opponentId: string;
+    opponentUsername: string;
+    result: string;
+    playerScore: number;
+    opponentScore: number;
     date: DateTime
 };
